Refactor road importation - Move from road to upload
