+++ backend/src/roads/road.service.ts
@@ -2,7 +2,6 @@ import { Injectable } from '@nestjs/common';
 import { InjectConnection, Knex } from 'nestjs-knex';
 import { IWay, Measurement, Way } from '../tables';
 import { LatLng, MeasurementType, OSMWayId, Road } from '../models';
-import { getOSMWaysInARoad } from './osm';
 import { constructNavigationTable, findLongestBranch } from './roads';
 
 @Injectable()
@@ -10,124 +9,6 @@ export class RoadService {
   constructor(
     @InjectConnection('group-d') private readonly knex_groupd: Knex,
   ) {}
-
-  /**
-   * Insert a way in the database.
-   * @param way the way to insert
-   * @returns the inserted way (in an array)
-   *
-   * @author Kerbourc'h
-   */
-  async insertWay(way: IWay) {
-    const { section_geom, ...rest } = way;
-
-    return Way(this.knex_groupd)
-      .insert({
-        ...rest,
-        section_geom: this.knex_groupd.raw('ST_GeomFromGeoJSON(?)', [
-          JSON.stringify(section_geom),
-        ]),
-      })
-      .onConflict('osm_id')
-      .merge() // Force to update because returning doesn't work with ignore()
-      .returning([
-        'id',
-        'way_name',
-        'osm_id',
-        'node_start',
-        'node_end',
-        'length',
-        this.knex_groupd.raw(
-          'ST_AsGeoJSON(section_geom)::json as section_geom',
-        ),
-        'isoneway',
-      ]);
-  }
-
-  /**
-   * Get or create the ways corresponding to the OSMWayIds.
-   * You should definitely use this function instead of getOrCreateWay if you
-   * have multiple ways to get.
-   * This executes the queries sequentially to avoid useless queries to OSM
-   * and the database. This is especially useful when the ways are in the same road.
-   *
-   * @param wayIds
-   *
-   * @author Kerbourc'h
-   */
-  async getOrCreateWays(wayIds: OSMWayId[]) {
-    let tasks: any = wayIds.map((id, index) =>
-      index === 0 ? this.getOrCreateWay(id).then((value: IWay) => [value]) : id,
-    );
-
-    return await tasks.reduce(async (cur: Promise<IWay[]>, next: string) => {
-      return cur.then(async (value: IWay[]) => {
-        if (((value.length / tasks.length) * 100) % 10 === 0)
-          console.info(
-            'Importation status',
-            `${(value.length / tasks.length) * 100}%`,
-          );
-
-        return [...value, await this.getOrCreateWay(next)];
-      });
-    });
-  }
-
-  /**
-   * Will get the Way corresponding to the OSMWayId either from the database if
-   * available or from OSM and insert the whole road containing it in the database.
-   *
-   * @param OSMWayId the osm id of a way in the road
-   *
-   * @author Kerbourc'h
-   */
-  async getOrCreateWay(OSMWayId: OSMWayId): Promise<IWay> {
-    const way = await Way(this.knex_groupd)
-      .select(
-        'id',
-        'way_name',
-        'osm_id',
-        'node_start',
-        'node_end',
-        'length',
-        this.knex_groupd.raw(
-          'ST_AsGeoJSON(section_geom)::json as section_geom',
-        ),
-        'isoneway',
-      )
-      .where('osm_id', OSMWayId)
-      .limit(1);
-
-    // If the way doesn't exist in the database, fetch the road from OSM
-    // and insert it in the database. Then return the way.
-    if (way.length === 0) {
-      console.info(
-        `Way ${OSMWayId} doesn't exist in the database, fetching road from OSM.`,
-      );
-      return await getOSMWaysInARoad(OSMWayId, async (data): Promise<IWay> => {
-        if (data.length > 0) {
-          console.info(
-            `Inserting road containing ${data.length} ways in the database.`,
-          );
-          return await Promise.all(data.map((way) => this.insertWay(way))).then(
-            async (result): Promise<IWay> => {
-              const ways = result.flat();
-              console.debug(
-                "Road's OSM Ids:",
-                ways.map((way) => way.osm_id).join(', '),
-              );
-              return ways.filter((way) => way.osm_id === OSMWayId)[0];
-            },
-          );
-        } else {
-          console.error('No way found in the road. Check the OSM Id.');
-          return null;
-        }
-      });
-    }
-
-    return way[0];
-  }
 
   /**
    * Query the geometry, node_start and node_end of each way and structure the data properly
